face_recognition_service/requirements.txt
opencv-python
face_recognition
flask
cloudinary
python-dotenv
requests
pymongo
dnspython
bcrypt
python-jose[cryptography] 